requirements.txt
django
django-localflavor